fix(slider): update cursor for disabled state and fix stories
