+++ core/components/atoms/slider/__stories__/variants/DiscreteSlider.story.jsx
@@ -2,13 +2,13 @@ import * as React from 'react';
 import { Slider } from '@/index';
 
 // CSF format story
-export const disctereSlider = () => {
+export const discreteSlider = () => {
   return (
-    <Slider min={1} max={10} label="Discrete Slider" stepSize={1} labelStepSize={1} defaultValue={4} className="mt-2" />
+    <Slider min={1} max={10} label="Discrete Slider" stepSize={1} labelStepSize={1} defaultValue={4} className="my-8" />
   );
 };
 
 export default {
-  title: 'Components/Slider/Variants/Disctere Slider',
+  title: 'Components/Slider/Variants/Discrete Slider',
   component: Slider,
 };
