+++ core/components/atoms/slider/__stories__/variants/FreeSlider.story.jsx
@@ -4,7 +4,7 @@ import { Slider } from '@/index';
 // CSF format story
 export const freeSlider = () => {
   return (
-    <Slider min={1} max={10} label="Free Slider" stepSize={0.1} labelStepSize={1} defaultValue={4} className="mt-2" />
+    <Slider min={1} max={10} label="Free Slider" stepSize={0.1} labelStepSize={1} defaultValue={4} className="my-8" />
   );
 };
