+++ docs/src/pages/components/slider/usage.mdx
@@ -14,7 +14,7 @@ Based on usage, sliders are of two types -
 
 The default slider can be used whenever selection of a single value is required out of a range.
 
-<Preview name='components-slider-variants-uncontrolled-slider--uncontrolled-slider'/>
+<Preview name='components-slider-types-uncontrolled-slider--uncontrolled-slider'/>
 
 #### Range slider
 
@@ -31,7 +31,7 @@ Based on selection, sliders have two variants -
 In the case of discrete slier, while dragging slider knob will snap from one tick mark to the other.
 No values between two tick marks can be selected.
 
-<Preview name='components-slider-variants-disctere-slider--disctere-slider'/>
+<Preview name='components-slider-variants-discrete-slider--discrete-slider'/>
 
 #### Free slider
